pyproject.toml
[project]
name = "bloggingplatformapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
]
